appium_test/requirements.txt
# Python dependencies for Appium testing
appium-python-client==3.1.0
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
